component cleanup done, renamed login form to match components folder. keeping ui folder lowercase for shadcn deps checks
